+++ CHANGELOG.md
@@ -6,7 +6,12 @@
 
 - Publicly exposed interface `CustomHeaderSecurity` is renamed to `HeaderSecurity`;
 - The default value for `wrongMethodBehavior` config option is changed to `405`;
-- Minimum version of `compression` (optional peer dependency) is `1.8.0` (it supports Brotli).
+- Minimum version of `compression` (optional peer dependency) is `1.8.0` (it supports Brotli);
+- The following public methods of `Endpoint` intended for internal usage became properties:
+  - `getDescription()` —> `description` and `shortDescription`;
+  - `getSchema()` —> `inputSchema` and `outputSchema`;
+  - `getMethods()` —> `methods`; `getTags()` —> `tags`; `getScopes()` —> `scopes`;
+  - `getSecurity()` —> `security`; `getRequestType()` —> `requestType`
 
 ## Version 22